Move to top-level otlp mod
