+++ src/exporters/datadog/transform/cloud.rs
@@ -3,7 +3,7 @@
 use crate::semconv;
 use opentelemetry_semantic_conventions::attribute;
 use std::str::FromStr;
-use crate::exporters::datadog::transform::cv_attributes::{ConvertedAttrMap, ConvertedAttrValue};
+use crate::otlp::cvattr::{ConvertedAttrMap, ConvertedAttrValue};
 
 #[allow(dead_code)]
 const EC2_TAG_PREFIX: &str = "ec2.tag.";
