+++ src/exporters/datadog/transform/k8s.rs
@@ -2,7 +2,7 @@
 
 use crate::exporters::datadog::transform::cloud;
 use opentelemetry_semantic_conventions::attribute;
-use crate::exporters::datadog::transform::cv_attributes::ConvertedAttrMap;
+use crate::otlp::cvattr::ConvertedAttrMap;
 
 pub(crate) fn hostname_from_attrs(attrs: &ConvertedAttrMap) -> Option<String> {
     attrs
