+++ src/exporters/datadog/transform/otel_mapping/attributes.rs
@@ -4,7 +4,7 @@ use crate::semconv::containers::CONTAINER_IMAGE_TAG;
 use opentelemetry_semantic_conventions::attribute;
 use std::collections::HashMap;
 use std::sync::LazyLock;
-use crate::exporters::datadog::transform::cv_attributes::ConvertedAttrMap;
+use crate::otlp::cvattr::ConvertedAttrMap;
 
 pub const CUSTOM_CONTAINER_TAG_PREFIX: &str = "datadog.container.tag.";
